oforha_db/requirements.txt
pymongo[srv]
python-dotenv
pytest 